requirements.txt
pillow
autopep8